Try to refactor the unit tests
